Refactored the world time. Now it is stored in two values - WorldAge (only incremented, plugins cannot change) and TimeOfDay (plugins can change). Since sub-tick precision is needed in Tick(), we store it both as number of seconds (double) and number of ticks (Int64) is calculated off of that.
git-svn-id: http://mc-server.googlecode.com/svn/trunk@1022 0a769ca7-a7f5-676a-18bf-c427514a06d6

+++ source/Pawn.h
@@ -58,7 +58,7 @@ protected:
 	MetaData m_MetaData;
 
 	double m_LastPosX, m_LastPosY, m_LastPosZ;
-	float m_TimeLastTeleportPacket;
+	Int64 m_TimeLastTeleportPacket;  // In ticks
 }; //tolua_export
